+++ tests/modules/test_alerts.py
@@ -69,9 +69,7 @@ class TestAlertsModule(object):
         )
         alert_module = AlertsModule(mock_microservice_client_factory)
         alert_module.resolve(self._alert_ids)
-        mock_alerts_client.resolve.assert_called_once_with(
-            self._alert_ids, tenant_id=None, reason=None
-        )
+        mock_alerts_client.resolve.assert_called_once_with(self._alert_ids, reason=None)
 
     def test_alerts_module_calls_reopen_with_expected_value(
         self,
@@ -84,6 +82,4 @@ class TestAlertsModule(object):
         )
         alert_module = AlertsModule(mock_microservice_client_factory)
         alert_module.reopen(self._alert_ids)
-        mock_alerts_client.reopen.assert_called_once_with(
-            self._alert_ids, tenant_id=None, reason=None
-        )
+        mock_alerts_client.reopen.assert_called_once_with(self._alert_ids, reason=None)